LayoutTests/imported/w3c/web-platform-tests/workers/constructors/SharedWorker/interface-objects-expected.txt


FAIL Test if interface objects exist in a shared worker assert_equals: expected "These were missing: " but got "These were missing: Worker"
